Fixed charging DeadbeatCrossbowItem

+++ src/client/java/me/gv0id/arbalests/client/render/entity/state/EndCrystalProjectileEntityRenderState.java
@@ -6,5 +6,5 @@ import net.minecraft.entity.Entity;
 public class EndCrystalProjectileEntityRenderState extends EndCrystalEntityRenderState {
     public boolean invisible = false;
     public static Entity owner;
-    public static int fuse;
+    public static float fuse;
 }
